refactor: move removing unused images to shared supabase slice
